Tensorflow_ML-DL/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
dm-tree==0.1.9
docstring_parser==0.17.0
einops==0.8.1
etils==1.13.0
flatbuffers==25.9.23
fsspec==2025.10.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
immutabledict==4.2.2
importlib_resources==6.5.2
keras==3.12.0
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.3.4
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
promise==2.3
protobuf==6.33.0
psutil==7.1.3
pyarrow==22.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rich==14.2.0
simple-parsing==0.1.7
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-datasets==4.9.9
tensorflow-metadata==1.17.2
termcolor==3.2.0
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.0
zipp==3.23.0
